+++ utility/src/main/java/StatisticsHandler.java
@@ -1,6 +1,6 @@
 public interface StatisticsHandler {
     void handleStr(String strVal);
-    void handleInt(int intVal);
+    void handleInt(Long intVal);
     void handleFloat(float floatVal);
     String statistics();
 }
